Rewritten to fix several quirks:
* aP on the first parameter will not include the comma for consistency
(because if it does, the comma is sometimes in front, and sometimes at
the end of the selection). There's a parameter that allows you to
configure the plugin to be more like the old behaviour.

* iP skips the surrounding spaces.

* aP includes the leading comma, and excludes the trailing spaces.

* The dependence on registers is removed. This allows very cool way to
exchange two parameters that I learned about about a couple of days
ago: move to first param, "diP" to cut it out, move to second param,
"viPp" to replace it, move to first param hole, "P" to paste. :)

* The dependence on marks is removed, so the user's marks are also preserved.

* I observed that the counts were not working after all (normal! "lyl
would reset the count variable, apparently). It should work now.

* I also included elements of lists and objects - basically, any
comma-separated lists - so if you do "viP" in foo(bar, [baz, quux]),
it will select "bar", "baz", "quux", and not "foo", "[baz", "quux]".
There is a slight problem if the braces are unbalanced, like foo([bar,
baz}) in that it completely ignores the differences between different
parentheses but I believe it's a minor issue.